+++ R/ccc.R
@@ -18,9 +18,7 @@
 #' or ICD-10-CM codes.  The \code{data.frame} passed to the function should be
 #' in wide format.
 #' @param id bare name of the column containing the patient id
-#' @param dx_cols column names with the diagnostic codes, use
-#' \code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).
-#' @param pc_cols the column(s) containing the procedure codes.  Use
+#' @param codes column names with the diagnostic codes, use
 #' \code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).
 #' @param icdv ICD version 9 or 10, defaults to 9
 #'
@@ -30,28 +28,8 @@
 #' @example examples/ccc.R
 #'
 #' @export
-ccc <- function(.data, id, dx_cols = NULL, pc_cols = NULL, icdv = 9) {
-
-  if (!is.null(dx_cols)) {
-    dxmat <- as.matrix(dplyr::select_(.data, .dots = dx_cols))
-  } else {
-    dxmat <- matrix("")
-  }
-
-  if (!is.null(pc_cols)) {
-    pcmat <- as.matrix(dplyr::select_(.data, .dots = pc_cols))
-  } else {
-    pcmat <- matrix("")
-  }
-
+ccc <- function(.data, id, codes, icdv = 9) {
+  cmat <- as.matrix(dplyr::select_(.data, .dots = codes))
   ids <- dplyr::select_(.data, .dots = lazyeval::interp( ~ i, i = substitute(id)))
-
-  Map(ccc_rcpp,
-      dx = split(dxmat, seq(1, nrow(dxmat))),
-      pc = split(pcmat, seq(1, nrow(pcmat))),
-      MoreArgs = list(version = icdv)) %>%
-  do.call(rbind, .) %>%
-  dplyr::as_data_frame() %>%
-  dplyr::bind_cols(ids, .)
+  dplyr::bind_cols(ids, ccc_rcpp(cmat, icdv))
 }
-