+++ src/RcppExports.cpp
@@ -17,15 +17,14 @@ BEGIN_RCPP
 END_RCPP
 }
 // ccc_rcpp
-Rcpp::IntegerVector ccc_rcpp(std::vector<std::string>& dx, std::vector<std::string>& pc, int version);
-RcppExport SEXP pccc_ccc_rcpp(SEXP dxSEXP, SEXP pcSEXP, SEXP versionSEXP) {
+Rcpp::DataFrame ccc_rcpp(Rcpp::CharacterMatrix& MAT, int version);
+RcppExport SEXP pccc_ccc_rcpp(SEXP MATSEXP, SEXP versionSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< std::vector<std::string>& >::type dx(dxSEXP);
-    Rcpp::traits::input_parameter< std::vector<std::string>& >::type pc(pcSEXP);
+    Rcpp::traits::input_parameter< Rcpp::CharacterMatrix& >::type MAT(MATSEXP);
     Rcpp::traits::input_parameter< int >::type version(versionSEXP);
-    rcpp_result_gen = Rcpp::wrap(ccc_rcpp(dx, pc, version));
+    rcpp_result_gen = Rcpp::wrap(ccc_rcpp(MAT, version));
     return rcpp_result_gen;
 END_RCPP
 }