+++ src/pccc.h
@@ -4,77 +4,58 @@
 #ifndef PCCC_H
 #define PCCC_H
 
-class codes {
+class ccc_codes {
   private:
+    // ICD version
     int version;
 
-    std::vector<std::string> dx_neuromusc; 
-    std::vector<std::string> dx_cvd; 
-    std::vector<std::string> dx_respiratory;
-    std::vector<std::string> dx_renal;
-    std::vector<std::string> dx_gi;
-    std::vector<std::string> dx_hemato_immu;
-    std::vector<std::string> dx_metabolic;
-    std::vector<std::string> dx_congeni_genetic;
-    std::vector<std::string> dx_malignancy;
-    std::vector<std::string> dx_neonatal;
-    std::vector<std::string> dx_tech_dep;
-    std::vector<std::string> dx_transplant;
-
-    std::vector<std::string> pc_neuromusc;
-    std::vector<std::string> pc_cvd;
-    std::vector<std::string> pc_respiratory;
-    std::vector<std::string> pc_renal;
-    std::vector<std::string> pc_gi;
-    std::vector<std::string> pc_hemato_immu;
-    std::vector<std::string> pc_metabolic;
-    //std::vector<std::string> pc_congeni_genetic;
-    std::vector<std::string> pc_malignancy;
-    //std::vector<std::string> pc_neonatal;
-    std::vector<std::string> pc_tech_dep;
-    std::vector<std::string> pc_transplant;
+    // Classificaitons
+    std::vector<std::string> codes_neuromusc; 
+    std::vector<std::string> codes_cvd; 
+    std::vector<std::string> codes_respiratory;
+    std::vector<std::string> codes_renal;
+    std::vector<std::string> codes_gi;
+    std::vector<std::string> codes_hemato_immu;
+    std::vector<std::string> codes_metabolic;
+    std::vector<std::string> codes_congeni_genetic;
+    std::vector<std::string> codes_malignancy;
+    std::vector<std::string> codes_neonatal;
+    std::vector<std::string> codes_tech_dep;
+    std::vector<std::string> codes_transplant; 
 
   public:
-    codes(int v);
-
+    // constructors
+    ccc_codes(int v);
+    
+    // accessors
     int get_version() { return version; };
 
-    int neuromusc(      std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int cvd(            std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int respiratory(    std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int renal(          std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int gi(             std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int hemato_immu(    std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int metabolic(      std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int congeni_genetic(std::vector<std::string>& dx);
-    int malignancy(     std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int neonatal(       std::vector<std::string>& dx);
-    int tech_dep(       std::vector<std::string>& dx, std::vector<std::string>& pc);
-    int transplant(     std::vector<std::string>& dx, std::vector<std::string>& pc);
-
-    std::vector<std::string> get_dx_neuromusc()       { return dx_neuromusc; };
-    std::vector<std::string> get_dx_cvd()             { return dx_cvd; };
-    std::vector<std::string> get_dx_respiratory()     { return dx_respiratory; };
-    std::vector<std::string> get_dx_renal()           { return dx_renal; };
-    std::vector<std::string> get_dx_gi()              { return dx_gi; };
-    std::vector<std::string> get_dx_hemato_immu()     { return dx_hemato_immu; };
-    std::vector<std::string> get_dx_metabolic()       { return dx_metabolic; };
-    std::vector<std::string> get_dx_congeni_genetic() { return dx_congeni_genetic; };
-    std::vector<std::string> get_dx_malignancy()      { return dx_malignancy; };
-    std::vector<std::string> get_dx_neonatal()        { return dx_neonatal; };
-    std::vector<std::string> get_dx_tech_dep()        { return dx_tech_dep; };
-    std::vector<std::string> get_dx_transplant()      { return dx_transplant; };
+    std::vector<std::string> get_neuromusc()       { return codes_neuromusc; };
+    std::vector<std::string> get_cvd()             { return codes_cvd; };
+    std::vector<std::string> get_respiratory()     { return codes_respiratory; };
+    std::vector<std::string> get_renal()           { return codes_renal; };
+    std::vector<std::string> get_gi()              { return codes_gi; };
+    std::vector<std::string> get_hemato_immu()     { return codes_hemato_immu; };
+    std::vector<std::string> get_metabolic()       { return codes_metabolic; };
+    std::vector<std::string> get_congeni_genetic() { return codes_congeni_genetic; };
+    std::vector<std::string> get_malignancy()      { return codes_malignancy; };
+    std::vector<std::string> get_neonatal()        { return codes_neonatal; };
+    std::vector<std::string> get_tech_dep()        { return codes_tech_dep; };
+    std::vector<std::string> get_transplant()      { return codes_transplant; };
                                                                            
-    std::vector<std::string> get_pc_neuromusc()       { return pc_neuromusc; };
-    std::vector<std::string> get_pc_cvd()             { return pc_cvd; };
-    std::vector<std::string> get_pc_respiratory()     { return pc_respiratory; };
-    std::vector<std::string> get_pc_renal()           { return pc_renal; };
-    std::vector<std::string> get_pc_gi()              { return pc_gi; };
-    std::vector<std::string> get_pc_hemato_immu()     { return pc_hemato_immu; };
-    std::vector<std::string> get_pc_metabolic()       { return pc_metabolic; };
-    std::vector<std::string> get_pc_malignancy()      { return pc_malignancy; };
-    std::vector<std::string> get_pc_tech_dep()        { return pc_tech_dep; };
-    std::vector<std::string> get_pc_transplant()      { return pc_transplant; };
+    // Check for a code
+    int neuromusc(      std::vector<std::string>& x);
+    int cvd(            std::vector<std::string>& x);
+    int respiratory(    std::vector<std::string>& x);
+    int renal(          std::vector<std::string>& x);
+    int gi(             std::vector<std::string>& x);
+    int hemato_immu(    std::vector<std::string>& x);
+    int metabolic(      std::vector<std::string>& x);
+    int congeni_genetic(std::vector<std::string>& x);
+    int malignancy(     std::vector<std::string>& x);
+    int neonatal(       std::vector<std::string>& x);
+    int tech_dep(       std::vector<std::string>& x);
+    int transplant(     std::vector<std::string>& x);
 };
 
 #endif