+++ NEWS.md
@@ -1,3 +1,9 @@
+# Version 0.1.1.9000
+
+## API Change
+* The `ccc` function only takes one entry for codes.  There is no need to have
+  the diagnostic and procedure codes inputted seperatly.
+
 # Version 0.1.1
 
 ## Bug Fixes